refactor(chain-network): move logic into sync/tip_poll.rs
