+++ services/chain/chain-network/src/sync/mod.rs
@@ -1,4 +1,5 @@
 pub mod config;
 pub mod orphan_handler;
+pub mod tip_poll;
 
 const LOG_TARGET: &str = "chain_network::service::sync";
